LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/semantics/the-button-element/command-and-commandfor/button-type-reflection.tentative-expected.txt


reset  submit  type=button  invalid  missing  missing with command only missing with commandfor only
reset submit type=button invalid missing missing with command only missing with commandfor only  reset submit type=button invalid missing missing with command only missing with commandfor only

PASS Button with id reset-in-form should reflect type correctly
PASS Button with id submit-in-form should reflect type correctly
PASS Button with id button-in-form should reflect type correctly
PASS Button with id invalid-in-form should reflect type correctly
FAIL Button with id missing-in-form should reflect type correctly assert_equals: type of missing-in-form should be button expected "button" but got "submit"
FAIL Button with id missing-in-form-command-only should reflect type correctly assert_equals: type of missing-in-form-command-only should be button expected "button" but got "submit"
FAIL Button with id missing-in-form-commandfor-only should reflect type correctly assert_equals: type of missing-in-form-commandfor-only should be button expected "button" but got "submit"
PASS Button with id reset-attr-form should reflect type correctly
PASS Button with id submit-attr-form should reflect type correctly
PASS Button with id button-attr-form should reflect type correctly
PASS Button with id invalid-attr-form should reflect type correctly
FAIL Button with id missing-attr-form should reflect type correctly assert_equals: type of missing-attr-form should be button expected "button" but got "submit"
FAIL Button with id missing-attr-form-command-only should reflect type correctly assert_equals: type of missing-attr-form-command-only should be button expected "button" but got "submit"
FAIL Button with id missing-attr-form-commandfor-only should reflect type correctly assert_equals: type of missing-attr-form-commandfor-only should be button expected "button" but got "submit"
PASS Button with id reset-outside-form should reflect type correctly
PASS Button with id submit-outside-form should reflect type correctly
PASS Button with id button-outside-form should reflect type correctly
PASS Button with id invalid-outside-form should reflect type correctly
FAIL Button with id missing-outside-form should reflect type correctly assert_equals: type of missing-outside-form should be button expected "button" but got "submit"
FAIL Button with id missing-outside-form-command-only should reflect type correctly assert_equals: type of missing-outside-form-command-only should be button expected "button" but got "submit"
FAIL Button with id missing-outside-form-commandfor-only should reflect type correctly assert_equals: type of missing-outside-form-commandfor-only should be button expected "button" but got "submit"
